refactor dashboard components: rename Driver to Collector, remove Stats and Notifications, and add new Dashboard and Bins components
