+++ app/dashboard/bins.tsx
@@ -1,0 +1,5 @@
+import Bins from '../../components/Bins';
+
+export default function BinsPage() {
+  return <Bins />;
+}
