+++ app/dashboard/collector.tsx
