+++ app/dashboard/dashboard.tsx
@@ -1,0 +1,5 @@
+import Dashboard from '../../components/Dashboard';
+
+export default function DashboardPage() {
+  return <Dashboard />;
+}
