--- app/dashboard/notifications.tsx
@@ -1,5 +1,0 @@
-import Notifications from '../../components/Notifications';
-
-export default function NotificationsPage() {
-  return <Notifications />;
-}
