--- app/dashboard/stats.tsx
@@ -1,10 +1,0 @@
-import Stats from '../../components/Stats';
-
-export default function StatsPage() {
-  return (
-    <div>
-      <h1>Welcome to the Stats Page</h1>
-      <Stats />
-    </div>
-  );
-}
