+++ components/Bins.tsx
@@ -1,0 +1,8 @@
+export default function Bins() {
+  return (
+    <div>
+      <h1 className="text-2xl font-bold mb-4">Bins Management</h1>
+      <p>Bins management content goes here...</p>
+    </div>
+  );
+}
