+++ components/Collector.tsx
