--- components/Notifications.tsx
@@ -1,8 +1,0 @@
-export default function Notifications() {
-  return (
-    <div>
-      <h1 className="text-2xl font-bold mb-4">Notifications</h1>
-      <p>Notifications content goes here...</p>
-    </div>
-  );
-}
